split DataElement into ABC and created new FlagElement class
