+++ test/museek/test_data_element.py
@@ -4,6 +4,7 @@ from unittest.mock import MagicMock, patch
 import numpy as np
 
 from museek.data_element import DataElement
+from museek.flag_element import FlagElement
 from museek.flag_list import FlagList
 
 
@@ -50,7 +51,7 @@ class TestDataElement(unittest.TestCase):
         expect = np.resize(np.arange(27), self.shape) / 2.
         np.testing.assert_array_equal(expect, divided._array)
 
-    @patch('museek.data_element.np')
+    @patch('museek.abstract_data_element.np')
     def test_getitem(self, mock_np):
         self.assertEqual(mock_np.squeeze.return_value, self.element[0, 1, 2])
         mock_np.squeeze.assert_called_once_with(5)
@@ -104,7 +105,7 @@ class TestDataElement(unittest.TestCase):
         flag_array[0, 0, 0] = True
         flag_array[1, 1, 1] = True
         flag_array[2, 2, 2] = True
-        flags = FlagList(flags=[DataElement(array=flag_array)])
+        flags = FlagList(flags=[FlagElement(array=flag_array)])
         mean = self.element.mean(axis=0, flags=flags)
         self.assertEqual(3, len(mean._array.shape))
         expect = np.asarray([[13.5, 10., 11.],
@@ -232,7 +233,7 @@ class TestDataElement(unittest.TestCase):
         flag_array[0, 0, 0] = True
         flag_array[1, 1, 1] = True
         flag_array[2, 2, 2] = True
-        flags = FlagList(flags=[DataElement(array=flag_array)])
+        flags = FlagList(flags=[FlagElement(array=flag_array)])
         mean = self.element._flagged_mean(axis=0, flags=flags)
         self.assertEqual(3, len(mean._array.shape))
         expect = np.asarray([[13.5, 10., 11.],
